+++ notifications/WingspanNotifications/Models/Shared/E974d9e9e5676123cf56d2fd858ece1c24be8be15b96a100963b7cb1afada329Status.cs
@@ -55,4 +55,5 @@ namespace WingspanNotifications.Models.Shared
             throw new Exception($"Unknown value {value} for enum E974d9e9e5676123cf56d2fd858ece1c24be8be15b96a100963b7cb1afada329Status");
         }
     }
+
 }