+++ notifications/WingspanNotifications/Models/Shared/FundingStatus.cs
@@ -53,4 +53,5 @@ namespace WingspanNotifications.Models.Shared
             throw new Exception($"Unknown value {value} for enum FundingStatus");
         }
     }
+
 }