+++ notifications/WingspanNotifications/Models/Shared/PlanType.cs
@@ -51,4 +51,5 @@ namespace WingspanNotifications.Models.Shared
             throw new Exception($"Unknown value {value} for enum PlanType");
         }
     }
+
 }