+++ notifications/WingspanNotifications/Models/Shared/ServiceState.cs
@@ -55,4 +55,5 @@ namespace WingspanNotifications.Models.Shared
             throw new Exception($"Unknown value {value} for enum ServiceState");
         }
     }
+
 }